requirements.txt
torch
pandas
numpy
wandb
scikit-learn
pyyaml